Add content_disposition as option in fog config
Paperclip did not have an option to change the content_disposition
header on uploads.

Needed to be able to upload Rackspace Cloud Files as
"Content-Disposition: attachment;" so that the app could prompt
downloads instead of opening files in browser. This could be
accomplished using send_file or send_data, but that requires
opening and downloading the file from CDN to appserver,
then allowing the client to download it from there.

more info here:
http://docs.rackspace.com/files/api/v1/cf-devguide/content/Enabling_Browser_Bypass_with_the_Content-Disposition_Header-d1e2219.html